Added observe share
