+++ app/src/main/java/com/koalatea/thehollidayinn/softwareengineeringdaily/podcast/PodcastDetailActivity.java
@@ -6,6 +6,7 @@ import android.content.Intent;
 import android.graphics.Color;
 import android.os.Build;
 import android.os.Bundle;
+import android.support.annotation.RequiresPermission;
 import android.support.v4.app.ShareCompat;
 import android.support.v4.content.ContextCompat;
 import android.support.v4.media.MediaBrowserCompat;
@@ -40,6 +41,7 @@ import com.koalatea.thehollidayinn.softwareengineeringdaily.repositories.UserRep
 import com.koalatea.thehollidayinn.softwareengineeringdaily.data.repositories.BookmarkDao;
 import com.koalatea.thehollidayinn.softwareengineeringdaily.downloads.MP3FileManager;
 import com.koalatea.thehollidayinn.softwareengineeringdaily.util.AlertUtil;
+import com.koalatea.thehollidayinn.softwareengineeringdaily.util.ReactiveUtil;
 import com.mikepenz.google_material_typeface_library.GoogleMaterial;
 import com.mikepenz.iconics.IconicsDrawable;
 import com.ohoussein.playpause.PlayPauseView;
@@ -319,16 +321,7 @@ public class PodcastDetailActivity extends PlaybackControllerActivity {
     mService.upVote(post.get_id())
             .subscribeOn(Schedulers.io())
             .observeOn(AndroidSchedulers.mainThread())
-            .subscribe(new DisposableObserver<Void>() {
-              @Override
-              public void onComplete() {}
-              @Override
-              public void onError(Throwable e) {
-                Log.v(TAG, e.toString());
-              }
-              @Override
-              public void onNext(Void posts) {}
-            });
+            .subscribe(ReactiveUtil.getEmptyObservable());
   }
 
   @OnClick(R.id.down_button)
@@ -361,18 +354,7 @@ public class PodcastDetailActivity extends PlaybackControllerActivity {
     mService.downVote(post.get_id())
       .subscribeOn(Schedulers.io())
       .observeOn(AndroidSchedulers.mainThread())
-      .subscribe(new DisposableObserver<Void>() {
-        @Override
-        public void onComplete() {}
-
-        @Override
-        public void onError(Throwable e) {
-          Log.v(TAG, e.toString());
-        }
-
-        @Override
-        public void onNext(Void posts) {}
-      });
+      .subscribe(ReactiveUtil.getEmptyObservable());
   }
 
   /* Downloads */
@@ -541,19 +523,7 @@ public class PodcastDetailActivity extends PlaybackControllerActivity {
     mService.addBookmark(post.get_id())
       .subscribeOn(Schedulers.io())
       .observeOn(AndroidSchedulers.mainThread())
-      .subscribe(new DisposableObserver<Void>() {
-        @Override
-        public void onComplete() {
-        }
-        @Override
-        public void onError(Throwable e) {
-          Log.v(TAG, e.toString());
-        }
-
-        @Override
-        public void onNext(Void posts) {
-        }
-      });
+      .subscribe(ReactiveUtil.getEmptyObservable());
 
     AppDatabase db = Room.databaseBuilder(getApplicationContext(), AppDatabase.class, "sed-db")
             .build();
